Support Venafi as a Service's new zone configuration name format
The old format (used by Venafi Cloud/Venafi DevOpsACCELERATE) was a UUID. Venafi as a Service's new format is 'app\api alias'. This requires VCert 0.11.

VCert 0.11 doesn't support the old UUID format anymore, so this vcert dependency upgrade is a breaking change, which is why we're upgrading our major version to 2.

We also downgrade our 'cryptography' package requirement to >=3.3.2 because VCert 0.11 requires <= 3.3.2 in order to maintain Python 2 compatibility[1]. This confuses 'setup.py install' and pip's pre-2020 resolver (which are inclined to install cryptography > 3.3.2), so we document this caveat and we work around it in the Dockerfile. See [2] for more info about pip's 2020 resolver.

Closes #4.

[1] https://github.com/Venafi/vcert-python/commit/ab64e57dca0bd99b3869c885fae717324b6fbd3b
[2] https://pip.pypa.io/en/latest/user_guide/#changes-to-the-pip-dependency-resolver-in-20-3-2020